requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.30.0
aiomysql==0.2.0
alembic==1.12.1

# Cache and Storage
redis==5.0.1
faiss-cpu==1.8.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Processing
numpy==1.26.4
pandas==2.2.3
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# ML Inference
tritonclient[http]==2.40.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0

# Environment
python-dotenv==1.0.0
